Correcciones EP2 - Parte 1

+++ Carrito-D/Carrito-D/Models/Compra.cs
@@ -21,9 +21,9 @@ namespace Carrito_D.Models
         public Carrito Carrito { get; set; }
 
         [Required(ErrorMessage = ErrorMsg.Requerido)]
-        [Range(50, float.MaxValue, ErrorMessage = ErrorMsg.Rango)]
+        [Range(50, decimal.MaxValue, ErrorMessage = ErrorMsg.Rango)]
         [DataType(DataType.Currency)]
-        public float Total { get; set; }
+        public decimal Total { get; set; }
 
         [Required(ErrorMessage = ErrorMsg.Requerido)]
         [DisplayFormat(ApplyFormatInEditMode = true, DataFormatString = "{0:dd/MM/yyyy HH:mm}")]
